Migrate add index button
